Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain
langchain-groq
groq
pillow==10.1.0
ultralytics==8.0.220
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
opencv-python==4.8.1.78
aiofiles==23.2.1
httpx==0.25.2